Split I2C into i2c.rs (new master API) and i2c_legacy.rs
Follow the crate's existing pattern (rmt/pcnt/timer) for legacy vs new
driver separation:

- Default: i2c.rs exposes I2cMasterBus/I2cMasterDevice (new API)
- With `i2c-legacy` feature: i2c_legacy.rs exposes the old I2cDriver,
  I2cSlaveDriver, and command-link based API

Shared types (I2c trait, I2cError, peripheral definitions) are
duplicated in both files following the same pattern as rmt_legacy.rs.
