+++ Cargo.toml
@@ -50,6 +50,10 @@ pcnt-legacy = []
 # - When enabled, the code for the legacy driver will be compiled.
 # - When disabled the code for the new driver will be compiled.
 timer-legacy = []
+# Similar to rmt-legacy
+# - When enabled, the code for the legacy I2C driver will be compiled.
+# - When disabled the code for the new I2C master bus/device driver will be compiled.
+i2c-legacy = []
 embedded-graphics = ["dep:embedded-graphics-core"]
 
 [dependencies]
